requirements.txt
fastapi~=0.101.0
pydantic~=1.10.0
returns~=0.21.0
uvicorn~=0.23.2
pymongo~=4.4.1
bunnet~=1.1.0
python-dotenv~=1.0.0
